ath10k-ct: Fix reported RSSI for wave-1.
wave-1 (988x in this case) does not appear to properly fill in the
secondary bandwidths signal properly, so the logic was summing up
garbage.  Disable the signal summing for wave-1 hardware,
which brings us back to similar to how upstream implemented
the code to begin with.  Wave-2 will still sum up the signal
as expected.

Signed-off-by: Ben Greear <greearb@candelatech.com>