mtk: mt76: mt7996: add sanity check for rx skb
Add sanity check for rx skb to prevent kernel crash
If the RXD is corrupted for an SKB, then drop it directly and record
the rx drop in band 0 (the band info in rxd cannot be trusted).
For mt7992, in testmode, an SKB with a corrupted RXD (the payload is
correct) might be sent by the firmware when fast calibration is enabled.
The actual root cause should be further investigated.

Signed-off-by: StanleyYP Wang <StanleyYP.Wang@mediatek.com>