wifi: mac80211: 0 out conn_settings at init
On older compilers (gcc-13 and older), this struct would be filled
with whatever was in the memory space before. This is fine for stock
because it is guarenteed the stock fields are assigned before usage.
For our CT kernels though, it is not always the case that we
interact with the conn_flags field of ieee80211_conn_settings,
which meant garbage would be left in the field. Zeroing out at
initialization ensures this won't be a problem any more.

Signed-off-by: Dylan Eskew <dylan.eskew@candelatech.com>