wifi: mt76: mt7925: Fix race for userspace ethtool stat querying
Without this, the length of the stats can change immediately after being
reported by the driver. This is because has_eht only is true after a
scan has been completed to ensure that the reg domain supports eht.
Instead, always report eht stats, even if they are not used by the
reg domain.

Signed-off-by: Rory Little <rory@candelatech.com>